full crudfunction  imageUP done boiler

+++ app/Http/Controllers/PostsController.php
@@ -3,11 +3,18 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 use DB;
 
 class PostsController extends Controller
 {
+
+    /*middel ware for auth */
+    public function __construct()
+    {
+        $this->middleware('auth',['except' => ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -43,12 +50,30 @@ class PostsController extends Controller
         $this->validate($request,[
 
             'title'=>'required',
-            'body'=>'required'
+            'body'=>'required',
+            'cover_image'=>'image|nullable|max:1999'
         ]);
+
+        //handle file upload
+        if($request->hasFile('cover_image')){
+            //how to get file name with the ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            //get just file ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore=$filename . '_' . time() . '.' . $extension;
+            //upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }else{
+            $fileNameToStore='noimage.jpg';
+        }
         $post=new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','post created');
@@ -75,6 +100,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $posts = Post::find($id);
+        //check user
+        if(auth()->user()->id !== $posts->user_id){
+            return redirect('/posts')->with('error',' UnAuth Page');
+        }
         return view('posts.edit')->with('posts',$posts);
 
     }
@@ -93,9 +122,28 @@ class PostsController extends Controller
             'title'=>'required',
             'body'=>'required'
         ]);
+
+          //handle file upload
+          if($request->hasFile('cover_image')){
+            //how to get file name with the ext
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //get just filename
+            $filename = pathinfo($filenameWithExt,PATHINFO_FILENAME);
+            //get just file ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename to store
+            $fileNameToStore=$filename . '_' . time() . '.' . $extension;
+            //upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images',$fileNameToStore);
+        }
+
+
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success','Post updated');
@@ -110,6 +158,16 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $posts = Post::find($id);
+        //check for correct users
+        if(auth()->user()->id !== $posts->user_id){
+            return redirect('/posts')->with('error',' UnAuth Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
+
         $posts->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
